Moved search-bar to utils/quicknav
